Refactor Resource Contexts for User Data
This enables refreshing of user or all instances of the resource.
